backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pydantic==2.10.2
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
email-validator==2.2.0
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.5.2
python-multipart==0.0.17
